\docType{data}
\name{ash30}
\alias{ash30}
\alias{ash30f}
\alias{ash30m}
\title{The ASH30 study}
\value{
  A data frame
}
\description{
  The ASH30 study comprises data on 195 people (84 males
  and 111 females) measured in 1980, when they were about
  twelve years old, and again in 2000, when they were about
  thirty-two. The project measured many variables but we
  will use only the height (h in metres), weight (w in
  kilograms) and the Body Mass Index (BMI): BMI = w / h^2.
}
\details{
  The \code{ash30} data set contains both males and
  females. The \code{ash30f} and \code{ash30m} data sets
  contain the data for female and male participents only.
}
\examples{
data(ash30)
}
\keyword{datasets}

