\docType{data}
\name{bodyweight}
\alias{bodyweight}
\alias{bodyweight1}
\alias{bodyweight2}
\alias{bodyweight3}
\alias{bodyweight4}
\alias{bodyweight5}
\alias{bodyweight6}
\title{The bodyweight data set}
\value{
  A data frame
}
\description{
  The average weight of the brain and body for sixty-two
  mammal species were recorded. We wish to predict the
  weight of the body using the brain weight. The first
  column is the subject index, the second is log of the
  brain weight and third is the log of the body weight.
}
\examples{
data(bodyweight)
}
\keyword{datasets}

