\docType{data}
\name{lathe}
\alias{lathe}
\title{The lathes data set}
\value{
  A data frame
}
\description{
  An experiment into the life of cutting tools for lathes
  was carried out. The lifetime is thought to depend on the
  the lathe speed and the type of material used for the
  tool. The lifetimes in hours, the type of material (A or
  B) and the the lathe speed in rpm were recorded for
  twenty tools.
}
\examples{
data(lathe)
}
\keyword{datasets}

