\docType{package}
\name{mas2316-package}
\alias{mas2316}
\alias{MAS2316}
\alias{mas2316-package}
\title{MAS2316}
\description{
  MAS2316
}
\details{
  Functions and datasets used in MAS2316: Introduction to
  regression and modelling
}
\author{
  \email{colin.gillespie@newcastle.ac.uk}
}
\seealso{
  See the Blackboard homepage for further details.
}
\keyword{package}

